% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharonBrainData.R
\docType{data}
\name{brain}
\alias{brain}
\title{Differential expression analysis using Sharon et al. (2020) data.}
\format{
A DESeq2 object containing results of differential expression analysis:
\describe{
\item{assays}{List of matrices with raw and normalized count data}
\item{rowRanges}{Genomic ranges for each gene}
\item{colData}{Data frame with sample metadata}
\item{design}{Formula representing the experimental design}
\item{dispersions}{Estimated dispersion values}
\item{results}{Data frame with differential expression results}
}
}
\source{
\url{doi:10.1016/j.cell.2019.05.004}
}
\usage{
brain
}
\description{
Expression data for regions of the mouse brain after inoculation of gut microbiota from typically developing humans. Data has been filtered to contain only sequences that were found in at least half of samples with a count of at least 10 reads.
}
\keyword{datasets}
