% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterProfilerGenesMouse.R
\name{getClusterProfilerGenesMouse}
\alias{getClusterProfilerGenesMouse}
\title{getClusterProfilerGenesMouse}
\usage{
getClusterProfilerGenesMouse(x, i)
}
\arguments{
\item{x}{The results of an over representation test - what genes are significantly up/down regulated - and their correlating KEGG categories. Typically the output of runClusterProfilerMouse}

\item{i}{A KEGG group from the dot plot. Case sensitive.}
}
\value{
Reports the gene names of genes found during a clusterProfiler analysis
}
\description{
Reports the gene names of genes found during a clusterProfiler analysis
}
