% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAcrossHostDonor.R
\name{plotAcrossHostDonor}
\alias{plotAcrossHostDonor}
\title{plotAcrossHostDonor}
\usage{
plotAcrossHostDonor(x)
}
\arguments{
\item{x}{A string containing a valid Mouse Genome Informatics (MGI) ID that is found in the dataset}
}
\value{
A figure plotting the gene expression of filtered TD-data from the Sharon et al. (2019) dataset across samples from a given host donor
}
\description{
Creates a figure plotting the gene expression using filtered TD-data from the Sharon et al. (2019) dataset across samples from a given host donor
}
