% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAcrossMice.R
\name{plotAcrossMice}
\alias{plotAcrossMice}
\title{plotAcrossMice}
\usage{
plotAcrossMice(x)
}
\arguments{
\item{x}{A string containing a valid Mouse Genome Informatics (MGI) ID that is found in the dataset}
}
\value{
A figure plotting the gene expression of filtered TD-data from the Sharon et al. (2019) dataset across mice
}
\description{
Creates a figure plotting the gene expression using filtered TD-data from the Sharon et al. (2019) dataset across mice.
}
