% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClusterProfilerMouse.R
\name{runClusterProfilerMouse}
\alias{runClusterProfilerMouse}
\title{runClusterProfilerMouse}
\usage{
runClusterProfilerMouse(x)
}
\arguments{
\item{x}{A data frame containing the results of a DESeq2 analysis}
}
\value{
The results of an over representation test - what genes are significantly up/down regulated - and their correlating KEGG categories
runClusterProfilerMouse()
}
\description{
Performs a clusterProfiler analysis on mouse data using KEGG categories from org.Mm.eg.db.
}
