% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharonBrainData-package.R
\docType{package}
\name{SharonBrainData-package}
\alias{SharonBrainData}
\alias{SharonBrainData-package}
\title{SharonBrainData: Data package for a subset of the data from Sharon et al. (2020)}
\description{
Contains data from Sharon et al. (2020) for the C-MOOR RNA-seq miniCURE. Only contains data from mice that were innoculated with typically developing human host mircobiomes. Data were pulled from recount3.
}
\author{
\strong{Maintainer}: Sayumi York \email{syork@ndm.edu}

}
\keyword{internal}
